requirements.txt
fastapi
uvicorn
tensorflow-cpu